REFACTOR: Move AppListViewModel.cloneApp() into IslandAppClones.
Hidden system apps treated as disabled are now handled in AppListViewModel before IslandAppClones.cloneApp().
